Add new EIP-4788 proof contracts, client script and utility script files.
This commit includes a set of scripts designed to aid the creation of proofs of block structures in the context of EIP-4788, as well as the import of contracts for proof verification. The scripts added include a client setup file and a utilities script, along with respective environment and configuration files. Moreover, the contract Verifier.sol allows for block root verification based on a given timestamp, while ValidatorVerifier.sol is to be filled with actual implementation details.