fix: let the create-validator command retrieve the loaded configuration.
Upon CLI bootstrapping, a PersistentPreRunE handler is added to the command
in order to, among other things, load the configuration in a viper instance and
store it into the command's context with a ViperContextKey key.

The create-validator command was using the wrong, but homonymous,
ViperContextKey for retrieving the viper configuration from the command
context.

As a result, the command implementation was not able to find some configuration
values which were silently used as empty strings that, eventually, would halt
the execution with unexpected errors.

This commit replaces the source package of the ViperContextKey key
from github.com/cosmos/cosmos-sdk/client
to github.com/berachain/beacon-kit/mod/cli/pkg/context