fix: kurtosis install script
There are two problems here:

1. On Ubuntu, the "command -v kurtosis &> /dev/null" fails to
detect if kurtosis is installed resulting in it trying to install
it again.

2. When extracting the kurtosis.tar.gz artifact it generates a
kurtosis binary in current folder. However we already have a
kurtosis/ folder in our repo so the extraction fails.

This commit fixes 1) so it works on any platform and addresses 2)
by removing the kurtosis automatic installation which should be
fine since its already a prerequisite in our kurtosis/README.md.