Remove all pre-proposal async events
This commit injects the BlockBuilder (the validator package) into
the cometbft.Service and calls it directly, removing the need for
NewSlot, BuiltBeaconBlock and BuiltSidecars events.

It was tricky to get DI working but eventually I found out that
injecting services through an interface with generic params or
return types is most likely not supported by the DI library. I
don't think this is documented anywhere, but you get a runtime
panic from DI once it can't assemble a dependecy based on the
interface definition. The solution here was to rely on only
concrete types, but this meant some additional struct generics
needed to be added to the cometbft.Service.