hack: Prevent node from starting up with invalid timeout_commit of 0
Since newest cometbft defaults to timeout_commit of 0 we always
default to that in our config.toml file as this code is generated
by cosmossdk init function handler.

To fix this properly I need to figure out how to pass in the config
as context value to the cobra.command as a viper instance.

Until then putting this error here as to preventing us forgetting
to set it to the correct time_commit (which I assume will be 1sec)