Initialize init command
Now that we have upgraded cosmossdk/cometbft dependencies it
becomes clear that we should improve the way we generate our
config files.

In newest cometbft version it set the default timeout_commit to
0 seconds instead of 1s it was before which which causes us to
produce blocks every 30-50us.

To address this quickly I implemented the InitCmd in our repo
instead of using its implmentation in cosmossdk.

Its almost the same except I check if timeout_commit is 0 and if
so set it to 1000ms.

But we should definetly fix the underlying issue of not using
proper defaults specific to our application.